#343 - package rename and adding logging that the extension is injected
