+++ dataset/src/main/scala/frameless/optimizer/Extension.scala
@@ -1,4 +1,4 @@
-package frameless.optimiser
+package frameless.optimizer
 
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.SparkSessionExtensions
@@ -6,5 +6,6 @@ import org.apache.spark.sql.SparkSessionExtensions
 class FramelessExtension extends ((SparkSessionExtensions) => Unit) with Logging {
   override def apply(extensions: SparkSessionExtensions): Unit = {
     extensions.injectOptimizerRule( _ => LiteralRule)
+    logInfo("injected OptimizerRule - frameless.optimizer.LiteralRule")
   }
 }
